requirements.txt
# ML Model Dependencies
torch
transformers
pillow
scikit-learn

# Agent Framework Dependencies
google-generativeai
python-dotenv

# Web Framework Dependencies
flask
flask-socketio
flask-cors
python-socketio
eventlet

# Utilities
colorama


